Keep the DDR shared in RCC
Do not clear RCC_DDRITFCFGR_DDRSHR because when it's clear the DDR
clock are linked to the status of the CPU1 instead of the status of the
system. This fix instability issues found when going in stop2.

Change-Id: Ie8998a76fd483b700804fb7606a4ada5e205bc7e
Signed-off-by: Thomas Bourgoin <thomas.bourgoin@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/stm32mp2-low-power/+/516785
Reviewed-by: Raphael GALLAIS-POU <raphael.gallais-pou@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Tested-by: Raphael GALLAIS-POU <raphael.gallais-pou@foss.st.com>
Reviewed-by: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
Domain-Review: Patrick DELAUNAY <patrick.delaunay@foss.st.com>